backend/requirements.txt
python-dotenv
transformers
langchain
langchain_openai
langchain_community
streamlit
fastapi
uvicorn
torch
PyPDF2
faiss-cpu
sentence-transformers
pycryptodome
python-multipart
sentencepiece
nltk
textstat
rouge-score